# Add 2-arg versions of findmax/min, argmax/min