--- a/base/exports.jl
+++ b/base/exports.jl
@@ -657,6 +657,7 @@ export
     isequal,
     ismutable,
     isless,
+    isgreater,
     ifelse,
     objectid,
     sizeof,