--- a/base/exports.jl
+++ b/base/exports.jl
@@ -657,7 +657,6 @@ export
     isequal,
     ismutable,
     isless,
-    isgreater,
     ifelse,
     objectid,
     sizeof,