--- a/base/reduce.jl
+++ b/base/reduce.jl
@@ -914,6 +914,12 @@ Values are compared with `isless`.
 
 # Examples
 ```jldoctest
+julia> argmax(abs, -10:5)
+-10
+
+julia> argmax(cos, 0:π/2:2π)
+0
+
 julia> argmax([8,0.1,-9,pi])
 1
 
@@ -945,6 +951,15 @@ Values are compared with `isgreater`.
 
 # Examples
 ```jldoctest
+julia> argmin(sign, -10:5)
+-10
+
+julia> argmin(x -> -x^3 + x^2 - 10, -5:5)
+5
+
+julia> argmin(acos, 0:0.1:1)
+1.0
+
 julia> argmin([8,0.1,-9,pi])
 3