… NEWS.md
… base/array.jl
… test/operators.jl
… test/reduce.jl
… test/reducedim.jl