Fixes #27613. Related: #27639, #27612, #34674.

Thanks to @tkf, @StefanKarpinski and @drewrobson for their assistance with
this PR.

I (and Stefan) wanted this to use partial order comparisons, but that would not be backwards compatible without a bit more faffing around. If there's interest, I might introduce another PR to explore that. That would introduce two partial orderings that bend the rules for these special cases: infinitesimals, unorderables (`NaN`) and `missing`.

I think reviewers should pay particular attention to the `isgreater` function I introduce. There are a lot of corner cases when comparing or ordering values and I'm not sure I've thought of them all. I'm fairly sure that it won't introduce any regressions in `findmin` or `argmin` (and it actually makes them work with `missing` where before they did not), but it may not be quite reliable enough to export.

Despite adding the `isgreater` function, I have not updated `min`, `minimum`, or the one-arg versions of `findmin` or `argmin` to use it. I can do that, but I wanted to get some review of what I've done first and possibly get this smaller change merged.