@cmcaine 2020-04-11
I don't think so. That's the branch for NaN and missing and I just want
those sorted the same as for isless.

On Sat, 11 Apr 2020, 17:09 Steven G. Johnson, <notifications@github.com>
wrote:

> *@stevengj* commented on this pull request.
> ------------------------------
>
> In base/operators.jl
> <https://github.com/JuliaLang/julia/pull/35316#discussion_r407080894>:
>
> > +"""
> +    isgreater(x, y)
> +
> +Test whether `x` is greater than `y`, according to a fixed total order.
> +`isgreater` is defined in terms of `isless`, but is not the opposite of that function.
> +
> +`isless` defines a fixed total order that ascends, while `isgreater` defines a
> +fixed total order that descends. Both order values that are normally unordered,
> +such as `NaN`, after all regular values and [`missing`](@ref) last. So for
> +`isless` these values are biggest and for `isgreater` these values are
> +smallest.
> +
> +# Implementation
> +Types should not usually implement this function. Instead, implement `isless`.
> +"""
> +isgreater(x, y) = _is_reflexive(x) && _is_reflexive(y) ? isless(y, x) : isless(x, y)
>
> Should that be !isless(x, y) in the final branch?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/JuliaLang/julia/pull/35316#pullrequestreview-391778905>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABNZA6K5FAOZUL7CN445ZWDRMCI37ANCNFSM4LXAJ2FQ>
> .
>