@StefanKarpinski 2020-05-02
Yes, this would be nice to have. What remains to be done here? Is this ready to go?