@cmcaine 2020-05-02
I paused on this because I found that the existing methods use an optimisation to avoid calling `isless` for values that are normally incomparable (so the first incomparable element will be returned, regardless of what `isless` wants).

The new definitions don't have that optimisation and are a bit slower, even when it is guaranteed that there are zero incomparables in the element type of the iterable. I wanted to try and work out why that was, but didn't get around to diving into the `isless` code to find out.

I can just not do that and then I think I only need to change the docstrings a bit to make sure we're giving a loose enough guarantee for the weird values.