@tkf 2020-06-14
I summarized the current issues around reducers in the "`findmax` family" in #36287.  It'd be nice if we can use a more mathematically rigorous and hence optimizable definition of them.