@cmcaine 2020-10-20
The only remaining issues are:

1. This is a minor change in semantics.
2. The current implementations are (or were, last I checked) faster due to the thing about skipping isless I mentioned.
3. Should probably replace my isgreater() with something like tkf's which is simpler.

Need to think about if the ::Colon and reducedim methods need to be done now or not. If they're not done they may be inconsistent.

Similarly, min, and maximum will be a bit inconsistent with findmax if not updated here, but they're not consistent now, so it probably doesn't really matter. Though would be good to fix.