@StefanKarpinski 2020-10-22
While I can see the motivation for `isgreater` for helping to implement these kinds of functions, but I'm hesitant about exporting it. My worry is that people will be tempted to modify or use `isgreater` independent of `isless`, which might be problematic. The legitimate use of `isgreater` is to impelment the same kind of logic as this implements for `findmax/min` and `argmax/min`, so maybe that's ok. If that's all people use it for, ok. But if people define `isless` and then think they also need to define `isgreater` to match or that they can define `isgreater` not to match... that's what worries me.